% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AboutApp.R
\name{about_app}
\alias{about_app}
\title{Refactored output$AboutSection server code into function 'about_app'}
\usage{
about_app()
}
\value{
This function will not return anything within the console. Its purpose is to demonstrate the refactoring of the previous 'output$AboutSection' server code into a new 'about_app' function.
}
\description{
shows user About section within the shiny application 'Coronavirus in Australia'
}
