% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the shiny application 'Coronavirus in Australia}
\usage{
launch_app()
}
\value{
Executing this function will return the 'Coronavirus in Australia' shiny application.
}
\description{
This function launches the shiny application 'Coronavirus in Australia'. The app allows the user to interact with Australia's coronavirus data.
}
