% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oz_covid_data}
\alias{oz_covid_data}
\title{Australia's Coronavirus data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 192 rows and 7 columns.
}
\source{
The dataset is a processed version of Rami Krispin's c'oronavirus' dataset from the 'coronavirus package' which can be found here: https://github.com/RamiKrispin/coronavirus
}
\usage{
oz_covid_data
}
\description{
The dataset contains a date variable and corresponding variables for total 'Daily Confirmed' coronavirus cases, 'Daily Death' related to coronavirus, 'Daily Recovered' coronavirus cases, 'Cumulative Confirmed' coronavirus cases, 'Cumulative Deaths' related to coronavirus and 'Cumulative Recovered' coronavirus cases. The 'date' column is of type <date> and the remaining columns are of type <int>.
}
\details{
This dataset contains Australia's daily and cumulative coronavirus data for 2020
}
\keyword{datasets}
