% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateSelectInput.R
\name{stateSelectInput}
\alias{stateSelectInput}
\title{Refactored 'selectInput' ui code}
\usage{
stateSelectInput()
}
\value{
This function will not return anything within the console. Its purpose is to demonstrate the refactoring of the previous 'selectInput' ui code into a new 'stateSelectInput' function.
}
\description{
Allows user to select state/territory of choice within shiny app to generate graph of confirmed coronavirus cases over time.
}
