% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vicEventRenderText.R
\name{vicEventRenderText}
\alias{vicEventRenderText}
\title{Refactored server-side output$VicDeathParagraph code into new function 'vicEventRenderText' function.}
\usage{
vicEventRenderText()
}
\value{
This function will not return anything meaningful for the end-user within the console. Its purpose is to demonstrate the refactoring of the previous 'output$VicDeathParagraph' server code into a new 'vicEventRenderText()' function. The code for the function may be used as a template by the user to include their own 'event_data' server code within a different shiny app.
}
\description{
Within the shiny application, this function generates a sentence describing the date and number of deaths that occurred in Victoria related to coronavirus on that date by asking the user to click on any point on the Victorian coronavirus-related deaths graphic. There is no need to include an argument within the function when executing the function.
}
