% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oz_covid_data}
\alias{oz_covid_data}
\title{Australia's Coronavirus data}
\format{
a tibble with 192 observations and 7 variables:
\itemize{
\item \strong{date}: the date on a particular day
\item \strong{'Daily Confirmed'}: the daily confirmed coronavirus cases in Australia on that given date
\item \strong{'Daily Death'}: the daily confirmed coronavirus deaths in Australia on that given date
\item \strong{'Daily Recovered'}: the daily confirmed coronavirus drecoveries in Australia on that given date
\item \strong{'Cumulative Confirmed'}: the cumulative confirmed coronavirus cases in Australia on that given date
\item \strong{'Cumulative Deaths'}: the cumulative confirmed coronavirus deaths in Australia on that given date
\item \strong{'Cumulative Recovered'}: the cumulative confirmed coronavirus recoveries in Australia on that given date
}
}
\source{
The dataset is a processed version of Rami Krispin's 'coronavirus' dataset from the 'coronavirus package' which can be found here: https://github.com/RamiKrispin/coronavirus
}
\usage{
oz_covid_data
}
\description{
This dataset contains Australia's daily and cumulative coronavirus data for 2020
}
\keyword{datasets}
