% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateSelectInput.R
\name{stateSelectInput}
\alias{stateSelectInput}
\title{Refactored 'selectInput' ui code into 'stateSelectInput' function}
\usage{
stateSelectInput()
}
\value{
This function will not return anything meaningful for the end-user within the console. Its purpose is to demonstrate the refactoring of the previous 'selectInput' ui code into a new 'stateSelectInput' function. The code for the function may be used as a template by the user to include their own 'selectInput' UI within a different shiny app.
}
\description{
Allows user to select state/territory of choice within shiny app to generate graph of confirmed coronavirus cases over time.
}
